+++ inc/Engine/Cache/ServiceProvider.php
@@ -11,7 +11,7 @@ use WP_Rocket\Engine\Preload\Database\Queries\Cache as CacheQuery;
 use WP_Rocket\Engine\Cache\Config\ConfigSubscriber;
 use WP_Rocket\Engine\Cache\UrlValidation\{ TaxonomySubscriber, PostSubscriber };
 use WP_Rocket\Engine\Cache\Abilities\{
-	PurgeUrl as PurgeUrlAbility,
+	ClearUrlCache as ClearUrlCacheAbility,
 	Subscriber as CacheAbilitiesSubscriber
 };
 
@@ -36,7 +36,7 @@ class ServiceProvider extends AbstractServiceProvider {
 		'cache_config',
 		'taxonomy_subscriber',
 		'post_subscriber',
-		'cache_abilities_purge_url',
+		'cache_abilities_clear_url_cache',
 		'cache_abilities_subscriber',
 	];
 
@@ -110,11 +110,11 @@ class ServiceProvider extends AbstractServiceProvider {
 			);
 		$this->getContainer()->addShared( 'taxonomy_subscriber', TaxonomySubscriber::class );
 		$this->getContainer()->addShared( 'post_subscriber', PostSubscriber::class );
-		$this->getContainer()->add( 'cache_abilities_purge_url', PurgeUrlAbility::class );
+		$this->getContainer()->add( 'cache_abilities_clear_url_cache', ClearUrlCacheAbility::class );
 		$this->getContainer()->addShared( 'cache_abilities_subscriber', CacheAbilitiesSubscriber::class )
 			->addArguments(
 				[
-					'cache_abilities_purge_url',
+					'cache_abilities_clear_url_cache',
 					'abilities_context',
 				]
 			);
